Migrate utils/deepEqual to TS
